+++ src/Application/Posts/Commands/CreatePost/CreatePostCommand.cs
@@ -1,25 +1,22 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using MediatR;
-using Microsoft.AspNetCore.Http;
 using SharedPhotoAlbum.Application.Common.Interfaces;
 using SharedPhotoAlbum.Application.Posts.Queries.GetPosts;
 using SharedPhotoAlbum.Domain.Entities;
-using SharedPhotoAlbum.Domain.Enums;
 
 namespace SharedPhotoAlbum.Application.Posts.Commands.CreatePost
 {
-    public class CreatePostCommand : IRequest<int>
+    public class CreatePostCommand : IRequest<long>
     {
         public IList<StoredMediaDto> StoredMedia { get; set; }
         public string Text { get; set; }
     }
 
-    public class CreatePostCommandHandler : IRequestHandler<CreatePostCommand, int>
+    public class CreatePostCommandHandler : IRequestHandler<CreatePostCommand, long>
     {
         private readonly IApplicationDbContext _dbContext;
 
@@ -28,7 +25,7 @@ namespace SharedPhotoAlbum.Application.Posts.Commands.CreatePost
             _dbContext = dbContext ?? throw new ArgumentNullException(nameof(dbContext));
         }
         
-        public async Task<int> Handle(CreatePostCommand request, CancellationToken cancellationToken)
+        public async Task<long> Handle(CreatePostCommand request, CancellationToken cancellationToken)
         {
             var media = request.StoredMedia.Select(_ => new StoredMedia
             {
